requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-ollama
langchain-text-splitters
faiss-cpu
beautifulsoup4
lxml
python-dotenv